+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/composable/ComposableSend.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.keyscreen.impl.composable.actions
+package com.flipperdevices.keyscreen.emulate.composable
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.layout.Column
@@ -20,48 +20,63 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
-import com.flipperdevices.keyscreen.impl.R
-import com.flipperdevices.keyscreen.impl.composable.actions.common.ComposableActionDisable
-import com.flipperdevices.keyscreen.impl.composable.actions.common.ComposableActionFlipper
-import com.flipperdevices.keyscreen.impl.model.EmulateButtonState
-import com.flipperdevices.keyscreen.impl.viewmodel.EmulateViewModel
+import com.flipperdevices.keyscreen.emulate.R
+import com.flipperdevices.keyscreen.emulate.composable.common.ComposableActionDisable
+import com.flipperdevices.keyscreen.emulate.composable.common.ComposableActionFlipper
+import com.flipperdevices.keyscreen.emulate.composable.common.ComposableActionLoading
+import com.flipperdevices.keyscreen.emulate.composable.common.ComposableAlreadyOpenedAppDialog
+import com.flipperdevices.keyscreen.emulate.model.DisableButtonReason
+import com.flipperdevices.keyscreen.emulate.model.EmulateButtonState
+import com.flipperdevices.keyscreen.emulate.viewmodel.EmulateViewModel
 import tangle.viewmodel.compose.tangleViewModel
 
 private const val SEND_BUTTON_SCALE = 1.06f
 
 @Composable
 fun ComposableSend(modifier: Modifier = Modifier, flipperKey: FlipperKey) {
-    val flipperDeviceViewModel = tangleViewModel<EmulateViewModel>()
-    val emulateButtonState by flipperDeviceViewModel.getEmulateButtonStateFlow().collectAsState()
+    val emulateViewModel = tangleViewModel<EmulateViewModel>()
+    val emulateButtonState by emulateViewModel.getEmulateButtonStateFlow().collectAsState()
 
     if (!flipperKey.synchronized) {
         ComposableActionDisable(
             modifier = modifier,
             textId = R.string.keyscreen_send,
-            iconId = DesignSystem.drawable.ic_send
+            iconId = DesignSystem.drawable.ic_send,
+            reason = DisableButtonReason.NOT_SYNCHRONIZED
         )
         return
     }
 
+    if (emulateButtonState == EmulateButtonState.AppAlreadyOpenDialog) {
+        ComposableAlreadyOpenedAppDialog(emulateViewModel::closeDialog)
+    }
+
     when (emulateButtonState) {
-        EmulateButtonState.DISABLED -> ComposableActionDisable(
+        is EmulateButtonState.Disabled -> {
+            ComposableActionDisable(
+                modifier = modifier,
+                textId = R.string.keyscreen_send,
+                iconId = DesignSystem.drawable.ic_send,
+                reason = (emulateButtonState as EmulateButtonState.Disabled).reason
+            )
+        }
+        is EmulateButtonState.Inactive,
+        is EmulateButtonState.Active -> ComposableSendInternal(
             modifier = modifier,
-            textId = R.string.keyscreen_send,
-            iconId = DesignSystem.drawable.ic_send
-        )
-        EmulateButtonState.INACTIVE,
-        EmulateButtonState.ACTIVE -> ComposableSendInternal(
-            modifier = modifier,
-            isAction = emulateButtonState == EmulateButtonState.ACTIVE,
+            isAction = emulateButtonState is EmulateButtonState.Active,
             onTap = {
-                flipperDeviceViewModel.onSinglePress(flipperKey)
+                emulateViewModel.onSinglePress(flipperKey)
             },
             onLongPressStart = {
-                flipperDeviceViewModel.onStartEmulate(flipperKey)
+                emulateViewModel.onStartEmulate(flipperKey)
             },
             onLongPressEnd = {
-                flipperDeviceViewModel.onStopEmulate()
+                emulateViewModel.onStopEmulate()
             }
+        )
+        is EmulateButtonState.Loading -> ComposableActionLoading(
+            modifier = modifier,
+            loadingState = (emulateButtonState as EmulateButtonState.Loading).state
         )
     }
 }
@@ -95,7 +110,6 @@ private fun ComposableSendInternal(
     ) {
         if (!isAction) {
             Text(
-                modifier = Modifier.padding(top = 4.dp),
                 text = stringResource(id = R.string.keyscreen_sending_desc),
                 style = LocalTypography.current.subtitleM12,
                 color = LocalPallet.current.text20,