+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/composable/common/ComposableActionFlipper.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.keyscreen.impl.composable.actions.common
+package com.flipperdevices.keyscreen.emulate.composable.common
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.RawRes
@@ -9,25 +9,15 @@ import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.unit.dp
 
 @Composable
+@Suppress("LongParameterList")
 fun ComposableActionFlipper(
     modifier: Modifier,
     color: Color,
@@ -57,30 +47,18 @@ fun ComposableActionFlipper(
         )
     ) else SolidColor(color)
 
-    Column(
+    ComposableEmulateButton(
         modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Row(
-            modifier = Modifier
-                .clip(RoundedCornerShape(12.dp))
-                .height(49.dp)
-                .fillMaxWidth()
-                .background(animColor)
-                .border(
-                    width = 2.dp,
-                    color = color,
-                    shape = RoundedCornerShape(12.dp)
-                ),
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
+        buttonModifier = Modifier,
+        buttonContent = {
             ComposableActionFlipperContent(
                 iconId = iconId,
                 textId = textId,
                 animId = if (isAction) animId else null
             )
-        }
-        content()
-    }
+        },
+        underButtonContent = content,
+        borderColor = color,
+        backgroundBrush = animColor
+    )
 }