+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/viewmodel/helpers/AlreadyOpenedAppException.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.keyscreen.emulate.viewmodel.helpers
+
+class AlreadyOpenedAppException : RuntimeException()