+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/model/EmulateButtonState.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.keyscreen.impl.model
-
-enum class EmulateButtonState {
-    DISABLED,
-    INACTIVE,
-    ACTIVE,
-}