+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Feature] Do not provide SubGhz provisioning on zero hardware region
 - [Feature] Restart RPC on fail RPC process response or flipper lags
 - [Feature] Prepare for Fdroid app store
+- [Feature] Handle errors while emulate/send
 - [FIX] Now search for devices by Mac address and by flipper name
 - [FIX] Subghz provisioning while geoip is null